refactor(agent): remove unused TLS certificate validation and related functions;
